Remove stateful 'victor'
Previously, we avoided calculating victor until the user guessed to
prevent cheating. We also were building around the idea of never
reloading the page.
This is no longer the case, and we simply calculate the victor as soon
as the page loads. Cheating isn't a concern since the easiest way to
cheat would just be to look up the moves, and simplifying the state both
avoids a race condition as well as makes the app faster.